Substitute "Point" with "PointService" class.
